requirements.txt
attrs==25.3.0
Automat==25.4.16
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
constantly==23.10.4
construct==2.10.70
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyparsing==3.2.3
pysnmp==7.1.20
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sympy==1.13.1
tensor==0.3.6
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
zope.interface==7.2
